make gentoo-based container work.
The change from ubuntu was "needed" due to gentoo not living
on quite the same plane of existance as Ubuntu, resulting
in gentoo ignoring the Ubuntu node due to iceccd "mismatch".
This mismatch could be fixed, but gcc and glibc mismatch make
that tentative rather moot.
And so, we have a gentoo container, for gentoo users.

A few kittens died in the process.